pyproject.toml
[build-system]
requires = ['hatchling']
build-backend = 'hatchling.build'

[project.scripts]
aj_bot = 'ajbot.__main__:_main'
aj_creds = 'ajbot.credentials:_main'

[tool.hatch.build.targets.pyinstaller]
require-runtime-dependencies = true
flags = ['--clean', '--onefile', '--noconfirm']
scriptname = 'src/ajbot/__main__.py'
specpath = './.spec'
zip = true

[project]
name = 'ajbot'
dynamic = ['version']
dependencies = ['discord.py',
                'sqlalchemy[asyncio]', 'aiomysql',
                'vbrpytools >= 3.8.0',
                'humanize',
                'dateparser',
                'thefuzz[speedup]',
                'pwinput',
               ]
description = 'Discord for personal use'
readme = 'README.md'
requires-python = '>=3.13'
license = {file = 'LICENSE'}
keywords = ['tools', 'discord','personal', 'mariaDB', 'bot']
authors = [
  { name='Vincent Brettnacher', email='vincent.brettnacher@veoneer.com' },
]
classifiers = [
    'Programming Language :: Python',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.13',
    'Operating System :: Microsoft :: Windows :: Windows 11',
    'Operating System :: POSIX :: Linux',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License'
]

[project.urls]
Documentation = 'https://github.com/vbrett/ajbot/blob/master/README.md'
source = 'https://github.com/vbrett/ajbot'
Repository = 'https://github.com/vbrett/ajbot.git'
Issues = 'https://github.com/vbrett/ajbot/issues'

[tool.hatch.version]
path = 'src/ajbot/__init__.py'